fix: move shared functionality to one location
