+++ apps/arclight/src/app/v2/taxonomies/route.ts
@@ -3,6 +3,8 @@ import { NextRequest } from 'next/server'
 
 import { getApolloClient } from '../../../lib/apolloClient'
 
+import { TaxonomyGroup, findBestMatchingName } from './lib'
+
 /* TODO: 
   querystring:
     apiKey
@@ -21,35 +23,6 @@ const GET_TAXONOMIES = graphql(`
     }
   }
 `)
-
-interface TaxonomyGroup {
-  terms: Record<
-    string,
-    {
-      label: string
-      metadataLanguageTag: string
-    }
-  >
-  _links: {
-    self: { href: string }
-    taxonomies: { href: string }
-  }
-}
-
-const findBestMatchingName = (
-  names: Array<{ label: string; languageCode: string }>,
-  preferredLanguages: string[]
-): { label: string; languageCode: string } => {
-  console.log('INSIDE findBestMatchingName')
-  console.log('names', names)
-  console.log('preferredLanguages', preferredLanguages)
-
-  for (const preferredLanguage of preferredLanguages) {
-    const match = names.find((name) => name.languageCode === preferredLanguage)
-    if (match !== undefined) return match
-  }
-  return names[0]
-}
 
 export async function GET(request: NextRequest): Promise<Response> {
   const { searchParams } = request.nextUrl
